Fix bug in choosing what day to run by default
The original code assumed that the %Map{} indicating which day to run
would be serialized in code-order, so the last line of code of that map
would be the day that it chose. But it actually sorts it in string order,
so once we get to day 10, it would just freeze up on day 9.

This fixes that by hard-coding the latest day.